Refactor commands.
