+++ generator/lib/download_command.dart
